% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{runApp}
\alias{runApp}
\title{Launch treetop application}
\usage{
runApp(...)
}
\arguments{
\item{...}{additional parameters from the \code{\link[shiny:runApp]{runApp}} function.}
}
\description{
This function launch the treetop application
}
\examples{
\dontrun{

# Launch treetop application
treetop::runApp(launch.browser = TRUE)

}
}
