% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchApp}
\alias{launchApp}
\title{Launch treetop application}
\usage{
launchApp(...)
}
\arguments{
\item{...}{additional parameters from the \code{\link[shiny:runApp]{runApp}} function.}
}
\description{
This function launch the treetop application
}
\examples{
\dontrun{

# Launch treetop application
treetop::launchApp(launch.browser = TRUE)

}
}
